Move get_mode() besides ForwardMode
This is more logical and ease its use by other modules.
