+++ src/decomon/metrics/loss.py
@@ -6,9 +6,8 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.layers import Lambda
 
 from decomon.layers.activations import softmax as softmax_
-from decomon.layers.core import DecomonLayer, ForwardMode
+from decomon.layers.core import DecomonLayer, ForwardMode, get_mode
 from decomon.models.models import DecomonModel
-from decomon.models.utils import get_mode
 from decomon.utils import ConvexDomainType, get_lower, get_upper, set_mode
